WGIT-5 Update Interface
